[MyS3CatsIOClient] extracted
